zeedit: new repository separate from `lbrytools`
The `zeedit` program uses the generic `lbrytools` library.
Originally, `zeedit` was hosted in the same `lbrytools` repository,
but since it is not really part of that library,
starting with this commit it is moved into its own repository.
All previous commits are shared with the `lbrytools` repository
so that it is possible to inspect
the history of development of `zeedit`.

Now `zeedit` can be installed by itself, and `lbrytools`
can be installed as a dependency, for example, as a Git submodule,
or as an external system library.

Simultaneously, the `zeedit` code will be removed completely
from the `lbrytools` repository, so that this library
contains generic code, and not an end-user application.